I changed a few things, but I mostly added the Point class and adjusted all of the points in the files to reflect that. It really changes very little, content-wise, but I feel that it is more organized. You may also noticed that I got rid of all of the massive includes in the include.h file (which is now gone, replaced by constants.h). That wasn't proper.
And finally, I changed jumping very slightly, so that Jump is now a method of the Character class. Also, instead of the isJumping boolean, I set it up as a flag with the enum value FLAG_JUMPING under flagList, so that it is the same as all of the other flags used for the class.