Although, at first glance, it will appear as though nothing has changed, I believe that I have greatly increased the power of the Animation class. Animations are now handled independently of frame rate, so that each new clip is only changed after so many milliseconds. And, each animation can have it's own update length, so that more intricate animations and more simple ones can coexist.
Another thing changed, instead of accessing the SetVelocity function to move left and right, I have some new functions MoveLeft, MoveRight, and StopMove, which do exactly what you would expect. They also set the animations up for themselves, which makes the Update methods of all of the dynamic objects much simpler and easier.

Of course, all of this framework is nice, and soon I'll actually have something to put in it. I'm thinking that on my next update, I will transfer the motion of the characters to be frame rate independent as well, since the amount to which they move left and right is still subject to the frame rate.

Also moved the animation creation section from the graphics class to the character class, since it makes more sense there.